- Improved MEItemInputBus and MEItemOutputBus, now have unlimited item cache capacity! - Cached item slots in MEItemInputBus and MEItemOutputBus can now not be interacted with via the GUI, only viewed. - MEItemInputBus can now use the scroll wheel to quickly modify the number of tagged items. - Improved the judgment method of the AE2EL mod.

+++ src/main/java/hellfirepvp/modularmachinery/client/gui/GuiMachineController.java
@@ -154,7 +154,7 @@ public class GuiMachineController extends GuiContainerBase<ContainerController> 
             }
         }
 
-        float usedTimeCache = TileMultiblockMachineController.usedTimeCache;
+        int usedTimeCache = TileMultiblockMachineController.usedTimeCache;
         float searchUsedTimeCache = TileMultiblockMachineController.searchUsedTimeCache;
         fr.drawStringWithShadow(String.format("Avg: %sμs/t (Search Avg: %sms)",
                         usedTimeCache,
